refactor: MemberStatus Enum을 Member 도메인 아래로 변경
