+++ src/main/java/umc/cockple/demo/domain/member/enums/MemberStatus.java
@@ -1,0 +1,6 @@
+package umc.cockple.demo.domain.member.enums;
+
+public enum MemberStatus {
+    ACTIVE,
+    INACTIVE // 회원 탈퇴시 INACTIVE
+}
